# Rework tree format + building process