--- a/boggle/breaker.py
+++ b/boggle/breaker.py
@@ -74,7 +74,9 @@ class HybridBreakDetails(BreakDetails):
     """Total nodes ever allocated (not necessarily all at once)"""
     tree_bytes: int
     """Number of bytes used by the initial orderly tree"""
-    tree_stats: TreeBuilderStats
+    n_paths: int
+    n_paths_uniq: int
+    tree_secs: list[float]
     total_bytes: int
     """Max bytes ever used while breaking"""
     n_bound: int
@@ -96,7 +98,7 @@ class HybridBreakDetails(BreakDetails):
         d["depth"] = counter_to_array(self.depth)
         d["bound_secs"] = float_dict_to_array(self.bound_secs)
         d["test_secs"] = round(self.test_secs, 5)
-        d["tree_stats"] = {k: round(v, 3) for k, v in d["tree_stats"].items()}
+        d["tree_secs"] = [round(v, 3) for v in d["tree_secs"]]
         return d
 
 
@@ -142,7 +144,7 @@ class HybridTreeBreaker:
             secs_by_level=defaultdict(float),
             bounds={},
             boards_to_test=0,
-            tree_stats=None,
+            tree_secs=[],
             init_nodes=0,
             depth=Counter(),
             total_nodes=0,
@@ -150,6 +152,8 @@ class HybridTreeBreaker:
             total_bytes=0,
             n_bound=0,
             n_force=0,
+            n_paths=0,
+            n_paths_uniq=0,
             max_multi=0,
             bound_secs=defaultdict(float),
             test_secs=0.0,
@@ -160,13 +164,9 @@ class HybridTreeBreaker:
         arena = self.etb.create_arena()
         tree = self.etb.build_tree(arena)
         ts = self.etb.get_stats()
-        self.details_.tree_stats = TreeBuilderStats(
-            collect_s=ts.collect_s,
-            sort_s=ts.sort_s,
-            build_s=ts.build_s,
-            n_paths=ts.n_paths,
-            n_uniq=ts.n_uniq,
-        )
+        self.details_.tree_secs = [ts.collect_s, ts.sort_s, ts.build_s]
+        self.details_.n_paths = ts.n_paths
+        self.details_.n_paths_uniq = ts.n_uniq
         num_nodes = arena.num_nodes()
         if self.log_breaker_progress:
             print(f"root {tree.bound=}, {num_nodes} nodes")