--- a/build.sh
+++ b/build.sh
@@ -13,6 +13,7 @@ c++ -Wall -shared -std=c++20 -fPIC -march=native \
     -Wno-sign-compare \
     -Wshadow \
     -Werror \
+    -DNDEBUG \
     -O3 \
     $(poetry run python -m pybind11 --includes | perl -pe 's/-I/-isystem /g') \
     cpp_boggle.cc trie.cc arena.cc eval_node.cc symmetry.cc \