… .inline-snapshot/external/8cd3c17dadcd3a9ba6c2de8137cc48a86601e6dd74040f2714fb9073d41ff130.txt
… .vscode/settings.json
… CLAUDE.md
… boggle/breaker_test.py
… boggle/equal_ranges_test.py
… boggle/eval_node_test.py
… boggle/orderly_tree_test.py
… boggle/packed5_array_test.py
… boggle/trie.py
… boggle/trie_test.py
… cpp/cpp_boggle.cc
… cpp/equal_ranges.h
… cpp/eval_node.cc
… cpp/eval_node.h
… cpp/orderly_tree_builder.h
… cpp/packed_array.h
… cpp/trie.cc
… cpp/trie.h
… testdata/3x4-2520743-1400.txt