Fixes #86 
Fixes #102 

Overall ~25% speedup on my Macbook, at least 20% reduction in RAM usage.

This includes the changes to `ChoiceNode` from #169.

The idea here is to build a flat list of all the ordered word paths, then sort+unique that to get the final list of paths to build the Sum/Choice tree. Contrast this with the previous approach, which added word paths one-by-one to the Sum/Choice tree.

The advantages of the new approach are:

1. By looking at the list, we can know in advance exactly how many children each SumNode / ChoiceNode will have. This saves memory, and also improves performance in subsequent steps thanks to greater memory locality.
2. There's no need for a hacky encoding system to de-dupe words (#140) We can use a standard sort + unique.
3. Building and sorting a list is considerably faster than building a tree. You still have to build the tree eventually, so this winds up being a wash.

The disadvantage is that we have two copies of the word paths live for a brief moment: the word list and the tree. The word list takes considerably less space than the tree (maybe 1/4 the space), and since the resulting tree uses less memory, this still winds up being a win. Some of this is due to interning childless SumNodes, which could also have been done with the old approach.

The encoding for word paths is pretty simple, 1 byte per cell and letter. This is wasteful — we only need 5 bits per cell and letter, not 8 — but it was hard to improve on this. Using a "packed array" required some care to ensure that `memcmp` still sorted lexicographically, and this made building the list and tree a bit slower. I also tried using the `used_ordered` bitmask to store which letters were active. This saved space but made sorting much, much slower. The current approach is simple, and the memory is only wasted temporarily. (Those approaches can be found in commits on this branch.)

Still to do:

- [x] Fix Python tests
- [x] Remove encoding
- [x] Remove AddWord?
- [x] Consolidate two branches in RangeToSum/ChoiceNode
- [x] Run tests with the 6-bit packed version
- [x] Record more tree building stats
- [x] Try more efficient version of `unique_ranges`
- [x] Look at [timing diffs per-board class](https://docs.google.com/spreadsheets/d/1uG-KFJFdOS7MQcibcdloTtQEQAdwror6VA0BO0AhkYQ/edit?gid=1019193302#gid=1019193302). Fairly consistent, but biggest win for mid-difficulty boards.
- [x] Try a `#define` macro-based versions of `CountWords` for perfect word list sizing.

Longer-term, the list representation might make it more feasible to save more memory by de-duping more nodes in the Sum/Choice tree.

`main`:

```
$ /usr/bin/time -l poetry run python -m boggle.break_all 'aeijou bcdfgmpqvwxz hklnrsty, corner:aeiosuy bcdfghjklmnpqrtvwxz' 3500 --size 44 --random_seed 2025 --max_boards 100 --switchover_score 6000 --output_base baseline-100.ndjson --num_threads 2
...
Broke 100 classes in 513.50s.
Found 0 breaking failure(s):

      514.40 real       975.23 user         9.22 sys
          2548678656  maximum resident set size
```

`remove-letter` (#169):

```
Broke 100 classes in 422.08s.
Found 0 breaking failure(s):

      423.08 real       780.00 user        12.15 sys
          2673852416  maximum resident set size
```

This PR:

```
Broke 100 classes in 375.84s.
Found 0 breaking failure(s):

      376.63 real       699.68 user         7.46 sys
          2218147840  maximum resident set size
```

This PR with `Packed6Array`:

```
Broke 100 classes in 435.10s.
Found 0 breaking failure(s):

      435.87 real       816.37 user         7.55 sys
          1818689536  maximum resident set size
```

Or for the best board's class:

```
(big-list) $ /usr/bin/time -l poetry run python -m boggle.break_all 'aeijou bcdfgmpqvwxz hklnrsty, corner:aeiosuy bcdfghjklmnpqrtvwxz' 3500 --size 44 --board_id 705707 --switchover_score 6000 --log_per_board_stats
Broke 1 classes in 702.99s.
Found 1 breaking failure(s):
plsteaiertnrsges
      703.72 real       699.03 user         3.03 sys
          3121561600  maximum resident set size
```

vs. main:

```
      794.46 real       777.88 user         7.64 sys
          4094017536  maximum resident set size
```

so 23% less RAM and ~11% faster.